# artifacts/api-server/src/routes/invoices.ts
import { Router, type IRouter } from "express";
import { eq, desc } from "drizzle-orm";
import { db, invoicesTable } from "@workspace/db";
import { requireAuth } from "../middlewares/auth";

const router: IRouter = Router();

router.get("/invoices", requireAuth, async (req, res): Promise<void> => {
  if (!req.user?.clinicId && req.user?.role !== "super_admin") {
    res.status(403).json({ error: "No clinic associated" });
    return;
  }

  const invoices = await db.select().from(invoicesTable)
    .where(eq(invoicesTable.clinicId, req.user.clinicId!))
    .orderBy(desc(invoicesTable.createdAt));

  res.json(invoices);
});

router.get("/invoices/:invoiceId", requireAuth, async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.invoiceId) ? req.params.invoiceId[0] : req.params.invoiceId;
  const invoiceId = parseInt(raw, 10);

  const [invoice] = await db.select().from(invoicesTable).where(eq(invoicesTable.id, invoiceId));

  if (!invoice) {
    res.status(404).json({ error: "Invoice not found" });
    return;
  }

  if (req.user?.role !== "super_admin" && req.user?.clinicId !== invoice.clinicId) {
    res.status(403).json({ error: "Access denied" });
    return;
  }

  res.json(invoice);
});

export default router;