# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic Digital Growth SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List doctors for a clinic
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/clinicId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"

    post:
      operationId: createDoctor
      tags: [doctors]
      summary: Add a doctor to a clinic
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/clinicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDoctorBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
  /clinics/{clinicId}/doctors/{doctorId}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get a doctor
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/clinicId"
        - $ref: "#/components/parameters/doctorId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update a doctor
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/clinicId"
        - $ref: "#/components/parameters/doctorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDoctorBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"

  # Appointments (public booking)
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices for the authenticated clinic
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
  /invoices/{invoiceId}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/invoiceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Admin
...
  /dashboard/today:
    get:
      operationId: getDashboardToday
      tags: [dashboard]
      summary: Get today's dashboard summary for authenticated doctor
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayDashboard"
  /dashboard/queue-summary:
    get:
      operationId: getQueueSummary
      tags: [dashboard]
      summary: Get real-time queue summary for the clinic
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueSummary"
  /dashboard/weekly-stats:
    get:
      operationId: getWeeklyStats
      tags: [dashboard]
      summary: Get weekly appointment stats for the clinic
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyStatPoint"
  /dashboard/patient-stats:
    get:
      operationId: getPatientStats
      tags: [dashboard]
      summary: Get total patient count and growth stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    clinicId:
      name: clinicId
      in: path
      required: true
      schema:
        type: integer
    doctorId:
      name: doctorId
      in: path
      required: true
      schema:
        type: integer
...
    invoiceId:
      name: invoiceId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth schemas
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialization:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        clinicId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - createdAt
    CreateDoctorBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        specialization:
          type: string
        qualification:
          type: string
        experience:
          type: string
        bio:
          type: string
      required:
        - name
        - email
        - phone
        - password
    UpdateDoctorBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        specialization:
          type: string
        qualification:
          type: string
        experience:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string

    # Appointment schemas
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        plan:
          type: string
          enum: [monthly, quarterly, yearly]
        status:
          type: string
          enum: [active, expired, cancelled, pending]
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        amount:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - plan
        - status
        - amount
        - createdAt
    SubscriptionStatus:
      type: object
      properties:
        isActive:
          type: boolean
        subscription:
          $ref: "#/components/schemas/Subscription"
        daysRemaining:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - isActive
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clinicId:
          type: integer
        paymentId:
          type: ["integer", "null"]
        subscriptionId:
          type: ["integer", "null"]
        amount:
          type: integer
        currency:
          type: string
        description:
          type: ["string", "null"]
        issuedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - clinicId
        - amount
        - currency
        - issuedAt
        - createdAt

    # Admin schemas
...
    TodayDashboard:
      type: object
      properties:
        totalAppointmentsToday:
          type: integer
        completedToday:
          type: integer
        waitingToday:
          type: integer
        currentToken:
          type: ["integer", "null"]
        nextToken:
          type: ["integer", "null"]
        totalPatientsAllTime:
          type: integer
        subscriptionStatus:
          $ref: "#/components/schemas/SubscriptionStatus"
      required:
        - totalAppointmentsToday
        - completedToday
        - waitingToday
        - totalPatientsAllTime
        - subscriptionStatus
    QueueSummary:
      type: object
      properties:
        date:
          type: string
        totalTokens:
          type: integer
        waitingCount:
          type: integer
        calledCount:
          type: integer
        completedCount:
          type: integer
        skippedCount:
          type: integer
        currentToken:
          type: ["integer", "null"]
        estimatedWaitMinutes:
          type: ["integer", "null"]
      required:
        - date
        - totalTokens
        - waitingCount
        - calledCount
        - completedCount
        - skippedCount
    WeeklyStatPoint:
      type: object
      properties:
        date:
          type: string
        appointments:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
      required:
        - date
        - appointments
        - completed
        - cancelled
    PatientStats:
      type: object
      properties:
        totalPatients:
          type: integer
        newThisMonth:
          type: integer
        newThisWeek:
          type: integer
        repeatPatients:
          type: integer
      required:
        - totalPatients
        - newThisMonth
        - newThisWeek
        - repeatPatients
...